fix(mypy): ensure atol/atol args are float by wrapping computed tolerances with float()

+++ agencitylab/core/multiscale.py
@@ -32,7 +32,7 @@ def _sample_step(axis: np.ndarray) -> float:
     axis = validate_axis(axis)
     diffs = np.diff(axis)
     step = float(diffs[0])
-    tolerance = np.finfo(float).eps * max(1.0, abs(step)) * 64.0
+    tolerance = float(np.finfo(float).eps * max(1.0, abs(step)) * 64.0)
     if not np.allclose(diffs, step, rtol=1e-10, atol=tolerance):
         raise ValueError("discrete multiscale extensions require uniformly sampled coordinates")
     return step
@@ -45,7 +45,7 @@ def _window_samples(window: float, axis: np.ndarray) -> int:
     if samples < 1:
         raise ValueError("w is smaller than one sampling interval")
     represented = samples * step
-    tolerance = max(np.finfo(float).eps * max(1.0, abs(window)) * 128.0, step * 1e-9)
+    tolerance = float(max(np.finfo(float).eps * max(1.0, abs(window)) * 128.0, step * 1e-9))
     if not np.isclose(represented, window, rtol=1e-9, atol=tolerance):
         raise ValueError("w must be an integer multiple of the sampling interval")
     return samples
